Apply negative margin to lists after <p>
As <p> tags may be commonly used to title lists following them, their
general bottom margin is too large to associate them with the list that
follows.

This commit applies a negative margin to lists to move them closer to
their title.
To avoid this behaviour, users will have to explicitly define the space
with a plain <br/><br/>.