Call Router#back when aborting Steps component
When the Steps component attempts to step one step back on its first
step, it attempts to lookup a negative-indexed step causing a client
error.

To fix this, by default, the Steps component calls Router#back when
attempting to step back on the first step element.
This behaviour can be overwritten by defining a custom beforeBack
handler on the specific step, allowing individual components to inject
their own abort logic if wanted.